+++ addon/locales/en/forms/new-platform-flexberry-service-bus-client-edit-form.js
@@ -6,5 +6,7 @@ export default {
   'dnsIdentity-caption': 'DNS Identity',
   'description-caption': 'Description',
   'subscriptions-caption': 'Subscriptions',
-  'sendingPermissions-caption': 'Sending permissions'
+  'sendingPermissions-caption': 'Sending permissions',
+  'connectionsLimit-caption': 'Connections limit',
+  'sequentialSent-caption': 'Sequential sent'
 };